move macro to test code where it is used
