+++ storage/src/global_state/shared.rs
@@ -1,6 +1,5 @@
 mod additive_map;
 mod correlation_id;
-pub(crate) mod macros;
 pub mod transform;
 
 pub use additive_map::AdditiveMap;
